requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.2
langchain-chroma>=0.0.1
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.18
sentence-transformers>=2.2.2
duckduckgo-search>=4.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.4.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
fastapi>=0.115.0
uvicorn>=0.34.0
boto3>=1.34.0
paramiko>=3.4.0
gunicorn>=21.2.0
langchain-mistralai>=0.2.7
google-search-results>=2.4.2
gradio>=5.0.0 